Fixing too tight-loop (temporarily)
Two problems with our set up. First, milliseconds is often 1ms which is
way too short of a time to be waiting for input asyncronously. So even
though we were just calling `web_pauseForMilliseconds` over and over the
execution stack got behind. So for now let's just always yield for 5
seconds and we can adjust it to something less noticble later.

Second we need to use the experimental emscripten_sleep_with_yield in
order to use the console because otherwise it gets blocked in the
execution loop... and so does the tab so we couldn't close it as our CPU
spiraled out of control.