Building out our asm-bridge & streamlining process a bit
Not a build tool, but I added copy.sh to move the stuff from the brogue
compilation over to the client file in the right place.

We successfully have exported the data from the asm.js into data that
the view can use. I just need to hook it up now.