Have the connections to the asm all hooked up.
The game crashes on input however which is not ideal. I suspect this has
to do with everything being an int being returned from EM_ASM_INT_V
while we need other data types. There may be a type mismatch that is
only occuring at runtime.

Doesn't help that the crash crashes so hard that there is no error
output.